Added initial version of the Visual Studio unit test project for CharLS
Visual Studio 2012 \ 2013 supports unit testing for C++. The advantage of the VS CPP test framework is that is integrates very well into the IDE (Test Explorer).
The purpose of the unit tests are to unit test the low level classes. The exisiting test project will be maintained and used to high level test CharLS on other platforms.