Added std namespace to sizeof_t and abs to improve C++ compatibility
Some pure C++ compilers only have sizeof_t defined in the std namespace. The C++ standard defines sizeof_t to be in the std namespace but allow vendors to also publish it in the global namespace (a common practice). The definition in both cases are identical. By using the correct std:: namespace scope compatibility with compilers like the Texas Instruments Code Composer Studio 6.0.0 is increased.