Restructure project files for Visual Studio 2015
Common project properties can be moved into a .prop file. This makes it easier to maintain it.
Note: The IDE does not like it when certain properties are moved to a globla .prop file
The temp output path can be set for c# projects, but this casues rebuilds inside VS2015.