Resolve spelling errors
Incorrect spelling makes the code looks sloppy. The spelling checker works the best on complete words that are seperated with underscores or lowercase-uppercase.
Hungarian nameing is also a blocker to validate variable and function naming.

Some real spelling errors are fixed: Compenent instead of Component, proces instead of process, Gratient instead of Gradient