Update CMake build files from 2.6 to modern CMake 3.9
The modern way of CMake build script has changed. To make CharLS more package friendly, update the overall CMake structure to modern CMake.
For the moment target CMake 3.9 as the minimum, as this version is available on the Travis CI build system.
Visual Studio 2017 15.9 Preview 4 is required on the Windows platform, but the Visual Studio solution file and MSBuild projects is on that platform the supported way to build CharLS.