Replace the CharLSNetTest.csproj format with the new .NET project format.
Note: it is not possible to use .NET Core for this unit test project. Only 1 dotnet.exe can be in the path, and while both the x64 and x86 version can be installed, only 1 can be active. This makes it impossible to test both the x64 and the x86 build.

Other changes:
- Remove JpegLSBitmapDecoder (unsupported by .NET Core)
- Rename solution name from Win32 to x86 to ensure CI build works.