Simplify encoding design by not creating temp copies of marker segments
The original design would create several JpegSegments instances and add them to the JpegStreamWriter. In a second step all segments would be serialized to the stream. There was not a real need to create these temporary instances. By moving the write functionality the JpegStreamWriter and the JpegStreamReader are now logical twins.