Encode the ABI version and architecture in the Windows DLL filename
To ensure that the correct DLL is loaded, encode the ABI version (currently 2) and the CPU architecture in the DLL file name.
The architecture was already encoded (but only for x64), the ABI was missing, but requested as there are projects that use open source modules link with v1 and others with v2.