Compile the C++ sample and charlstest with same hidden flag
It was observed that building the C++ sample on MacOS would trigger warnings when building CharLS as a static library. This was caused by a mixed mode of building with -fvisibility=hidden defined or not.
When CharLS is build as a shared library (DLL) this option doesn't leak, but when CharLS is build as a static library, this compile option is seen by the linker.
In essence -fvisibility=hidden should always be defiined to ensure that CharLS functions don't become "exported" again.