Add explicit namespace prefix for GCC 5.4 enum lookup
GCC 5.x has a problem with name lookups when an enum is used in a class that has a member function with the same name.
It is unclear in which version of GCC this has been fixed, but 7.4.0 and newer work fine.
GCC claims that 5.x is 100% C++14 compatible. Add extra namespace to support usage of GCC 5.x