Add the .def and .rc files to the charls CMakeLists.txt file
To ensure that when a DLL is build using CMake the functions are correctly exported and the DLL has a version number the .rc and .def file needed to included as resource.

Note: the x64 build doesn't need the explicit fucntion exports as there is only 1 way to export functions on the win-x64 platform.