Add C++ static analyzer annotations to the API
Both the MSVC and the gcc\clang C++ compiler provide additional attributes that can be used to assist the static analyzer build into these compilers to detect problems at compile time.
The problem is that these extension attributes are not cross compiler compatible, which requires additional macros to ensure that all supported compiler can compile the code.
The main focus of adding these attributes is on the external C API as in that area the most errors can be made.