Add a .clang-tidy configuration file
Clang-tidy is a tool to provide additional static analysis warningts. It makes sense to keep as many warnings as possible active. Warning that are not usefull in scope of this project are disabled with a rationale.
The latest version of Visual Studio 2019 can also use clang-tidy to perform background checking.