Add [[deprecated]] to the legacy API functions
For the next version the intent is to warn people not to use the legacy API anymore.
Copy the MSVC trick to introduce a define to suppress this warning by defining a preprocessor symbol: CHARLS_NO_DEPRECATED_WARNING

Where possible remove the usage of this API by the CharLS code itself. The unit test code that verify the legacy API is of course excluded.