Sample code: ensure that the resolution is set correctly in the SPIFF header
The sample code would not read all bmp_dib_header fields. This has been corrected.
It is often the case that the horizontal and vertical resolution are not defined in a BMP file (set to zero).
Added also an extra check that the width and height info is correctly set in a BMP file, this to prevent that memory buffer of 0 bytes is allocated.