Fixes #48: only do HP color transforms on 8 and 16 bit
The original  HP color transformation extensions only work on 8 bit RGB color images. It makes no sense to try to encode other bit per sample options as the HP implementation cannot decode it.
If the user wants to use color transformation it can upscale or downscale the pixels before encoding it with JPEG-LS.