Add the /utf-8 switch to the MSVC compiler options to force reading source as utf-8
All C and C++ source code is encoded as utf-8 without a signature (bom). The GCC and clang compiler will read by default source code files in that format.
The MSVC compiler will look for a signature (bom), if non is found then the default system codepace encoding is used.
/utf-8 ensure that the correct source and executing encoding is used and that conversions are validated.

As no special characters are used in the source code, this is only a technical correction.