Add additional asserts to the C sample application for clang-tidy
A width or height of zero would cause a malloc call with 0 bytes. Calling malloc with a size of 0 is implementation dependend and may return a non-null pointer which cannot be used but must be forwared to free.
Note: higher level runtime checks already prevent that encode_bmp_to_jpegls is called with these bad values.