Convert byte_stream_info into byte_span
The typical use case is to encode and decode to and from a memory buffer. For almost all practical cases this is the most efficient encoding/decoding mechanisme.
Decoding from a byte stream or encoding to a byte stream  with help of a callback function has been considered. There are however no significant advantages for the typical image sizes.