Extend jpegls_decoder constructor with option to read header
The typical use case is to create a  jpegls_decoder and then call read_header. This option is still possible but the constructor now accepts a boolean parameter to do this step automatically.
The frame_info and spiff_header are now cached in members, preventing the need to create these helper objects in client code.