Mask high input bits during encoding
When encoding images ensure that only the valid bits are used. For bit size 16 and 8 there is no change and memcpy will be used.
Encoding with masking is not measurable slower.

Setting explicit InlineFunctionExpansion for debug needs to be removed as it prevents debugging.  For release mode it is also not needed as MaxSpeed will already enable it.

Mask high input bits during encoding (interleave_mode::sample)

Mask high input bits during encoding