Use SFINAE on Container::value_type to prevent ambiguous overloads
STL containers have the ::value_type typedef. Use this as simple SFINAE check to
ensure API methods that should be used with an STL like container are only available
if a container is passed.
This prevents ambiguous overload problems.
There are more advanced solutions: is_container in Boost, Concepts in C++20. but for the C++ API this solution meets its purpose (and is supported on older compilers).