Do not enforce exactly c++14 standard when building charls
Per cmake documentation:

* https://cmake.org/cmake/help/git-stage/manual/cmake-compile-features.7.html#requiring-language-standards

For example, if C++ 11 features are used extensively in a project's
header files, then clients must use a compiler mode that is no less than
C++ 11. This can be requested with the code:

    target_compile_features(mylib PUBLIC cxx_std_11)

In this example, CMake will ensure the compiler is invoked in a mode of
at-least C++ 11 (or C++ 14, C++ 17, ...), adding flags such as
-std=gnu++11 if necessary. This applies to sources within mylib as well
as any dependents (that may include headers from mylib).