Use explicit ofstream.close() and replace static_cast with uniform initialization
The destructor of std::ofstream will close the stream, but not report any failures (by design).
Call close to trigger an exception when the close fails.

static_cast for constant values can be replaced with uniform initialization, this allows
to reduce the number of static_casts.

Change the bin+intermediate to be sub-directories of build for MSBuild builds.
This makes it easier to clean.