Resolve warning C4127, C6326, C26814
C4127 = conditional expression is constant
C6326 = Potential comparison of a constant with another constant
C26814 = The const variable 'RANGE' can be computed at compile-time

With C++17 if constexpr the const template constructions can be indicated.
Note: some code needs to be duplicated as constexpr cannot be mixed with a runtime check.