Move decode method from reader to decoder
The responsibility of the jpeg_stream_reader class is to parse the JPEG bit stream.
The actual decoding should not be done by the reader (same design as the writer).

Note: as the structure of a JPEG bit stream is very flexible the reader should still be the owner of the frame_info and parametes.