Simplify struct triplet (remove union)
The union inside the triplet struct is only used to assign different names to its members. The different names are only used in the HP color transforms and have no other purpose.
Simplify the struct by removing the union. In most uses cases the other algorithms operates on components and color space information is not applicable at all.