Enable code coverage in CI pipeline (#279)
The Azure CI Pipeline can analyze code coverage when running the unit tests. Enable this option. It provides an additional tab on the CI status report. On this tab a single code coverage number is displayed. The .coverage file can be downloaded and opened in Visual Studio to get more insights.
Only the debug DLL is code covered as it provided a good report. The release build has too many optimalizations to provide a report that is useful. 

Note 1: The unit test DLL needs to link with the CharLS .obj files to make it possible to unit test low level methods. This means however the the unit test DLL needs to be code covered and not the actual CharLS DLL.

Note 2: the Cobertura format was also tested, but it didn't provide additional details (only more warnings and complexity)
Enable code coverage in CI pipeline