Add check to prevent out-of-bounds read during decoding
Invalid input data may trigger an invalid read.
This scenaroi will be hit for invalid input data that completes
a scan but hasn't a marker start byte at the expected position.
The current algorithm will add additional bits to the cache, but this requires an additional check for the position.

Note: found with LibFuzzer