Use positive indexes to acces the line buffer during decoding (#313)
The line buffer is 2 pixels longer then required to make is easier to decode pixels on the edges. The start of the pointers was set to the first real pixel.
This caused the problem that negative indexes needed to be used in the code, which looks incorrect at first sight and makes the code harder to port.