Use positive indexes to acces the line buffer during encoding (#315)
The line buffer is 2 pixels longer then required to make is easier to encode pixels on the edges. The start of the pointers was set to the first real pixel.
This caused the problem that negative indexes needed to be used in the code, which looks incorrect at first sight and makes the code harder to port.
The same was already done for the decoding process.