Re-order the failure values of the enum jpegls_errc (breaking change)
Group the error IDs in errors that cannot be prevented and can be reported at runtime and in a set of errors that can be prevented.

Note 1: the error destination_too_small is considerd a runtime failure, but can in many cases be prevented.
Note 2: unexpected exceptions will now trigger a call to std::terminate (improved crash dumps)