Replace functors with function pointers (#324)
* Replace functors with function pointers during decoding

The functors were constructed on the heap and the actual implemenation was called using a virtual method call.
The main advantage of a functor (insert inline) could not be leveraged.
As a first step replace these functors with basic function pointers. It replaces a virtual call to a member function with a
function pointer call to a static method (no need to pass this pointer).

* Rename process_decoded_line.hpp in copy_from_line_buffer.hpp

* Replace functors with function pointers during encoding

The functors were constructed on the heap and the actual implementation was called using a virtual method call.
The main advantage of a functor (insert inline) could not be leveraged.
As a first step replace these functors with basic function pointers. It replaces a virtual call to a member function with a
function pointer call to a static method (no need to pass this pointer).

* Rename process_encoded_line.hpp to copy_to_line_buffer.hpp