Update the name of golomb_code to golomb_code_match and POD conversion
The main purpose of golomb_code_match is to see if there is a precomputed match which allows to just skip the bytes.
Rename the struct to make it clear it is not the actual glomb code itself.
Update the golomb_code_match struct back into a POD. The Windows x64 ABI cannot return non-POD structs in a single register even when the would fit.