Add support to retrieve height from a DNL marker segment. (#329)
DNL markers are part of the official JPEG-LS standard (but almost never used). Add support to scan the buffer for this
marker segment if the height is zero.
By design CharLS itself will never write DNL marker segment as it always knows the height of the image before it writes the SOF marker segment.