Prevent endless loop during decoding for invalid JPEG-LS input data
Fuzzing testing of the C# implementation found a scenario that would cause an endless loop during decoding. This defect was also present in the C++ implementation.
This condition was tested with an ASSERT. Using an ASSERT is fine when encoding, but an explicit check is needed when decoding as the input data is in that case untrusted.